test(vsg): uncomment some tests for loaders
