+++ packages/vue-docgen-api/src/script-handlers/displayNameHandler.ts
@@ -45,7 +45,7 @@ function getDeclaredConstantValue(prog: NodePath<bt.Program>, nameConstId: strin
 	const nodeDeclaratorArray = declarators.filter(
 		d => bt.isIdentifier(d.id) && d.id.name === nameConstId
 	)
-	const nodeDeclarator = nodeDeclaratorArray.length ? nodeDeclaratorArray[0] : undefined
+	const nodeDeclarator = nodeDeclaratorArray.length ? nodeDeclaratorArray[0] : null
 	return nodeDeclarator && nodeDeclarator.init && bt.isStringLiteral(nodeDeclarator.init)
 		? nodeDeclarator.init.value
 		: null
